requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
clarifai==10.0.1
clarifai-grpc==10.0.4
click==8.1.7
contextlib2==21.6.0
dataclasses-json==0.6.3
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
llama-index==0.9.34
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.8.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pypdf==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.14
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
schema==0.7.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.30.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
tritonclient==2.41.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
